requirements.txt
Django>=5.2.9
requests>=2.31.0
Pillow>=12.0.0
psycopg2-binary>=2.9.0
django-ratelimit>=4.1.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
django-redis>=5.2.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
channels>=4.0.0
channels-redis>=4.1.0
redis>=5.0.0


